dont look for views on sqlite

+++ packages/saltcorn-data/base-plugin/viewtemplates/list.js
@@ -49,13 +49,15 @@ const create_db_view = async (context) => {
 };
 
 const on_delete = async (table_id, viewname, { default_state }) => {
-  const sqlviews = (await get_existing_views()).map((v) => v.table_name);
-  const vnm = db.sqlsanitize(viewname);
-  const schema = db.getTenantSchemaPrefix();
-  if (sqlviews.includes(vnm))
-    await db.query(`drop view if exists ${schema}"${vnm}";`);
-  if (sqlviews.includes(vnm + "_sqlview"))
-    await db.query(`drop view if exists ${schema}"${vnm + "_sqlview"}";`);
+  if (!db.isSQLite) {
+    const sqlviews = (await get_existing_views()).map((v) => v.table_name);
+    const vnm = db.sqlsanitize(viewname);
+    const schema = db.getTenantSchemaPrefix();
+    if (sqlviews.includes(vnm))
+      await db.query(`drop view if exists ${schema}"${vnm}";`);
+    if (sqlviews.includes(vnm + "_sqlview"))
+      await db.query(`drop view if exists ${schema}"${vnm + "_sqlview"}";`);
+  }
 };
 
 const configuration_workflow = (req) =>
